# core/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Archura Core API",
    "version": "0.1.0",
    "description": "Identity, ownership, and component-session API for Archura. Production /v1 requests require the Worker service credential plus any endpoint-specific bearer credential."
  },
  "servers": [
    { "url": "/", "description": "Current core server" }
  ],
...
  "components": {
    "securitySchemes": {
      "ServiceAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Archura-Service-Authorization",
        "description": "Worker credential including the Bearer prefix, for example `Bearer svc_live_...`. Required in production and optional in direct local development."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "adm_..., sk_..., or sess_...",
        "description": "Platform-admin key, organization secret key, or account session token, according to the operation."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "example": "invalid_api_key" },
              "message": { "type": "string", "example": "The organization API key is invalid." },
              "request_id": { "type": "string", "example": "core/abc123-000001" }
            }
          }
        }
      },
...
      "CreateConfirmationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "owner@example.com" },
          "subdomain": { "type": "string", "pattern": "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$", "example": "mikes-bakery" }
        }
      },
      "ConfirmationCreated": {
        "type": "object",
        "required": ["id", "expires_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "expires_at": { "type": "string", "format": "date-time" },
          "confirm_url": { "type": "string", "format": "uri", "description": "Present only in development." }
        }
      },
      "VerifyConfirmationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["token"],
        "properties": {
          "token": { "type": "string", "pattern": "^cfm_(?:test|live)_", "example": "cfm_test_..." }
        }
      },
      "Account": {
        "type": "object",
        "required": ["id", "email"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" }
        }
      },
      "AccountSession": {
        "type": "object",
        "required": ["token", "expires_at"],
        "properties": {
          "token": { "type": "string", "pattern": "^sess_(?:test|live)_", "description": "Returned once and stored only as a hash." },
          "expires_at": { "type": "string", "format": "date-time" }
        }
      },
      "Organization": {
        "type": "object",
        "required": ["id", "name", "slug", "role", "is_default", "publishable_key", "sites", "created_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "role": { "type": "string", "enum": ["owner", "member"] },
          "is_default": { "type": "boolean" },
          "publishable_key": { "type": "string", "pattern": "^pk_(?:test|live)_" },
          "sites": { "type": "array", "items": { "type": "string" } },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "ConfirmationVerified": {
        "type": "object",
        "required": ["account", "organization", "subdomain", "session"],
        "properties": {
          "account": { "$ref": "#/components/schemas/Account" },
          "organization": { "$ref": "#/components/schemas/Organization" },
          "subdomain": { "type": ["string", "null"] },
          "session": { "$ref": "#/components/schemas/AccountSession" }
        }
      },
...
    },
    "responses": {
      "BadRequest": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unauthorized": { "description": "Missing or invalid service/admin/organization credential", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "Conflict": { "description": "Resource conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "RateLimited": { "description": "Rate limit exceeded", "headers": { "Retry-After": { "schema": { "type": "integer" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "InternalError": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unavailable": { "description": "Database unavailable", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    }
  },
  "paths": {
...
    "/v1/confirmations": {
      "post": {
        "tags": ["Confirmations"],
        "operationId": "createConfirmation",
        "summary": "Create a single-use email confirmation",
        "security": [{ "ServiceAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateConfirmationRequest" } } } },
        "responses": {
          "201": { "description": "Confirmation created", "headers": { "Cache-Control": { "schema": { "type": "string", "example": "no-store" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfirmationCreated" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        }
      }
    },
    "/v1/confirmations/verify": {
      "post": {
        "tags": ["Confirmations"],
        "operationId": "verifyConfirmation",
        "summary": "Consume a confirmation and create an account session",
        "security": [{ "ServiceAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VerifyConfirmationRequest" } } } },
        "responses": {
          "200": { "description": "Confirmation verified", "headers": { "Cache-Control": { "schema": { "type": "string", "example": "no-store" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfirmationVerified" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        }
      }
    },
...
  }
}